fix(airflowctl): fix pre-commit for command images
