+++ airflow-ctl/docs/images/command_hashes.txt
@@ -1,14 +1,14 @@
-main:d41d8cd98f00b204e9800998ecf8427e
-assets:d41d8cd98f00b204e9800998ecf8427e
-auth:d41d8cd98f00b204e9800998ecf8427e
-backfills:d41d8cd98f00b204e9800998ecf8427e
-config:d41d8cd98f00b204e9800998ecf8427e
-connections:d41d8cd98f00b204e9800998ecf8427e
-dag:d41d8cd98f00b204e9800998ecf8427e
-dagrun:d41d8cd98f00b204e9800998ecf8427e
-jobs:d41d8cd98f00b204e9800998ecf8427e
-pools:d41d8cd98f00b204e9800998ecf8427e
-providers:d41d8cd98f00b204e9800998ecf8427e
-variables:d41d8cd98f00b204e9800998ecf8427e
-version:d41d8cd98f00b204e9800998ecf8427e
-auth login:d41d8cd98f00b204e9800998ecf8427e
+main:3d89ac882e883c75b3bc51d71039f215
+assets:36718a4f36cac58a2063c0a094945b12
+auth:ef4122d3f5e4b2ac19cb0d3e12c8594b
+backfills:e0cba4448d576d1b53ea79d6dcdbe035
+config:cf4a23af9f09575e9ed50144d8d64f08
+connections:da4f6807ca2a265ed6d6e734b5355fe2
+dag:dab7c8aa1a62fa011b80bb7132bcc32a
+dagrun:7b3e06a3664cc7ceb18457b4c0895532
+jobs:806174e6c9511db669705279ed6a00b9
+pools:080b31538715f6cb625935ff3c943a46
+providers:d053e6f17ff271e1e08942378344d27b
+variables:d9001295d77adefbd68e389f6622b89a
+version:11da98f530c37754403a87151cbe2274
+auth login:348c25d49128b6007ac97dae2ef7563f
